[hub75] Apply review cleanups
- Drop the redundant rotation field from the Hub75Config struct: set_rotation()
  is called before setup() runs, so config_.rotation is already correct by the
  time setup() constructs the driver. Single source of truth.
- Simplify the DisplayRotation -> Hub75Rotation cast: drop the unnecessary
  intermediate static_cast<int>.
- Trim verbose comments down to the non-obvious WHY.

+++ esphome/components/hub75/hub75.cpp
@@ -196,12 +196,9 @@ void HOT HUB75Display::draw_pixels_at(int x_start, int y_start, int w, int h, co
 }
 
 void HUB75Display::set_rotation(display::DisplayRotation rotation) {
-  // Apply rotation at the hardware level via the hub75 driver. We deliberately do NOT
-  // call Display::set_rotation() — leaving the base class rotation_ at 0 prevents the
-  // base class draw helpers from rotating coordinates a second time on top of the
-  // driver's hardware rotation.
-  // DisplayRotation and Hub75Rotation share the same numeric values (0/90/180/270).
-  auto hw_rotation = static_cast<Hub75Rotation>(static_cast<int>(rotation));
+  // Don't call Display::set_rotation(): leaving the base class rotation_ at 0 prevents
+  // base-class draw helpers from rotating on top of the driver's hardware rotation.
+  auto hw_rotation = static_cast<Hub75Rotation>(rotation);
   if (this->driver_ != nullptr) {
     this->driver_->set_rotation(hw_rotation);
   } else {
